Put back

+++ backend/webapp/src/main/resources-filtered/docker/entrypoint.sh
@@ -13,14 +13,36 @@ if [ -s /run/secrets/*.context.xml ];
     echo "context.xml NOT FOUND, proceeding with default config"
 fi
 
-if [ -s /run/secrets/*.isis.properties ];
+
+if [ -s /run/secrets/*.shiro.ini ];
+  then
+    # Symlink shiro.ini.
+    ln -sf /run/secrets/*.shiro.ini /run/conf/shiro.ini
+    echo "FOUND shiro.ini."
+  else
+    echo "shiro.ini NOT FOUND, proceeding with default config"
+fi
+
+
+if [ -s /run/secrets/*.isis.properties ] && [ -s /run/secrets/*.logging.properties ];
   then
     # Symlink isis.properties.
     ln -sf /run/secrets/*.isis.properties /run/conf/isis.properties
-    echo "FOUND isis.properties"
+    ln -sf /run/secrets/*.logging.properties /run/conf/logging.properties
+    echo "FOUND isis.properties and logging.properties."
   else
-    echo "isis.properties NOT FOUND, proceeding with default config"
+    echo "isis.properties and/or logging.properties NOT FOUND, proceeding with default config"
 fi
+
+if [ -s /run/secrets/*.translations.po ];
+  then
+    # Symlink isis.properties.
+    ln -sf /run/secrets/*.translations.po /run/conf/translations.po
+    echo "FOUND translations.po."
+  else
+    echo "translations.po NOT FOUND, proceeding with default config"
+fi
+
 
 # Running Catalina
 echo "Starting Catalina:"
